requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
python-pptx
openai>=1.0.0
python-multipart
streamlit
requests
google-generativeai
